Modify stringLast aggregator to use a secondary time column while testing if it is the last one.
When Druid rollup is applied and stringLast is applied on rolledup records, since matching grouped records will have
same timestamp (__time col.) it will not return correct values. This fix lets users specify a different timestamp column
in stringlast spec, which can let them pick the last value based on that column.
New attribute in stringLast is 'tsColumn'

here is an example if we have records <__time, fooCol, lastModifiedTime> we can use lastModifiedTime column to
determine if the fooCol value is the last or not. This is useful when __time is same for multiple records
"aggregations": [
		      {
		"type": "stringLast",
		"name": "foo",
		"fieldName": "fooCol",
		"maxStringBytes": "10000",
		"tsColumn":"lastModifiedTime"
                    },